moved gather_on_r_new to common defined separate PropagatorNoneParaxial and PropagatorFresnel with steppers moved check_new_grid to PropagatorResamplingFresnel multiple style and doc string fixes updated the example
